modification du controller OrderItem

+++ Server/Repositories/OrderItemRepository/OrderItemRepository.cs
@@ -23,22 +23,30 @@ namespace Server.Repositories.OrderItemRepository
             return await _context.OrderItems.ToListAsync();
         }
 
-        public async Task AddAsync(OrderItem orderItem)
+        public async Task AddOrderItemAsync(OrderItem orderItem)
         {
             await _context.OrderItems.AddAsync(orderItem);
             await _context.SaveChangesAsync();
         }
 
-        public async Task UpdateAsync(OrderItem orderItem)
+        public async Task UpdateOrderItemAsync(OrderItem orderItem)
         {
             _context.OrderItems.Update(orderItem);
             await _context.SaveChangesAsync();
         }
 
-        public async Task DeleteAsync(OrderItem orderItem)
+        public async Task DeleteOrderItemByIdAsync(int orderId)
         {
-            _context.OrderItems.Remove(orderItem);
-            await _context.SaveChangesAsync();
+            var orderItem = await _context.OrderItems.FindAsync(orderId);
+            if (orderItem != null)
+            {
+                _context.OrderItems.Remove(orderItem);
+                await _context.SaveChangesAsync();
+            }
+            else
+            {
+                throw new Exception("OrderItem not found");
+            }
         }
     }
